+++ Mountain-UI-Companion/Utils/Constants.swift
@@ -14,4 +14,7 @@ enum Constants {
     static let github = "https://github.com/matthewfernst/Mountain-UI-Companion-App"
     static let buyMeCoffee = "https://www.buymeacoffee.com/matthewfernst"
     static let mountainUIDisplayGitub = "https://github.com/matthewfernst/Mountain-UI"
+    
+    // Bundle ID
+    static let bundleID = "com.matthewfernst.Mountain-UI-Companion"
 }
